Reading Current Position
Let it calibrate by spinning 360 degrees once. Then intentionally place the wheel at any desired position, then press enter in the serial monitor! It will read the position of the wheel!

There is a possibility of having multiple values, in that case we will consider the value which is nearest to the current position of the wheel(this part is not included, currently it shows all the values)!